pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "code-review-env"
version = "0.1.0"
description = "AI agent reviews pull requests: detects bugs, suggests fixes, assigns scores"
requires-python = ">=3.11"
dependencies = [
    "openenv-core",
    "pydantic>=2.7.1",
    "openai>=1.0.0",
]

[project.scripts]
server = "server.app:main"

[project.optional-dependencies]
server = [
    "fastapi==0.111.0",
    "uvicorn[standard]==0.30.1",
    "websockets==12.0",
]
